Fix instant death hits from attacker with weakness
The weakness potion effect could reduce an attackers damage points into negative numbers. Minecraft would treat the initial hit as 0 damage but save the last hurt damage as negative.

This negative lastHurt damage could increase the next damage attempt if the attacker attacked during the invulnerability period.

Minecraft allows damages during the invulnerability period if the next damage is more than the previous damage with the formula damage=(newDamage-lastDamage). If the lastDamage is negative then damage is added to the final damage which could instantly kill a player.